requirements.txt
streamlit==1.38.0 
psycopg2-binary==2.9.9 
pymupdf==1.24.9 
chromadb==0.5.5 
python-dotenv==1.0.1 
reportlab==4.2.2 
openai==1.44.1 
llama-index==0.11.7 
llama-index-embeddings-openai==0.2.4 
llama-index-vector-stores-chroma==0.2.0 
semanticscholar==0.8.3 
requests==2.32.3 
numpy==1.24.4 
tenacity==8.5.0